# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 🚀 Edu Management API
  description: Education Management System API collection
  version: 1.0.0
servers:
  - url: http://{{base_url}}{{api_version}}
  - url: http://{{base_url}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /users/change-password:
    post:
      tags:
        - 🔑 Auth
      summary: Change Password
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"current_password\": \"Password@123\",\n 
                \"new_password\": \"NewPassword@456\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /users/:
    get:
      tags:
        - 👥 Users
      summary: List Users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: limit
          in: query
          schema:
            type: integer
          example: '10'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /users/{id}:
    get:
      tags:
        - 👥 Users
      summary: Get User By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 👥 Users
      summary: Update User
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: '"{\n  \"is_active\": true,\n  \"role\": \"TEACHER\"\n}"'
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 👥 Users
      summary: Delete User
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...